Added number of connections to spike compression table entries, now enforcing spike compression when sorting connections, and building compression tables during connection creation to communicate target information during network creation

+++ nestkernel/stimulation_backend_mpi.cpp
@@ -1,3 +1,4 @@
+
 /*
  *  stimulation_backend_mpi.cpp
  *
@@ -27,7 +28,6 @@
 
 // Includes from nestkernel:
 #include "kernel_manager.h"
-#include "stimulation_backend.h"
 #include "stimulation_backend_mpi.h"
 #include "stimulation_device.h"
 
@@ -159,7 +159,7 @@ nest::StimulationBackendMPI::prepare()
   }
 
   // Add the id of device of the other thread in the vector_id_device and update the count of all device
-  for ( int id_thread = 0; id_thread < kernel().vp_manager.get_num_threads(); id_thread++ )
+  for ( size_t id_thread = 0; id_thread < kernel().vp_manager.get_num_threads(); id_thread++ )
   {
     // don't do it again for the master thread
     if ( id_thread != thread_id_master )
